[FEAT] Lua 스크립트 기반 청약 매물의 남은 토큰 수량 감소 처리 로직 추가

+++ woorepie/src/main/java/com/piehouse/woorepie/customer/repository/CustomerRepository.java
@@ -22,10 +22,10 @@ public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
     @Modifying
     @Query("UPDATE Customer c SET c.accountBalance = c.accountBalance - :amount WHERE c.customerId = :customerId AND c.accountBalance >= :amount")
-    long decreaseBalance(@Param("customerId") Long customerId, @Param("amount") long amount);
+    int decreaseBalance(@Param("customerId") Long customerId, @Param("amount") long amount);
 
     @Modifying
     @Query("UPDATE Customer c SET c.accountBalance = c.accountBalance + :amount WHERE c.customerId = :customerId")
-    long increaseBalance(@Param("customerId") Long customerId, @Param("amount") long amount);
+    int increaseBalance(@Param("customerId") Long customerId, @Param("amount") long amount);
 
 }
